+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.10.1 (tbd)
 
 - Improved trusted publishing
+- Added `fallbackLanguage` in `piral-translate`
 
 ## 1.10.0 (February 24, 2026)